refactor(units): extract units module from scattered logic across 3 files
Candidate 4 from architecture review: temperature conversion and unit
resolution logic was scattered across weather.js (celsiusToFahrenheit,
determineDisplayUnits, unitsForOpenMeteo), config.js
(processTemperatureOptions), and display.js (formatFeelsLike re-implementing
conversion). No home for unit logic — it leaked across three modules.

Extracted all unit functions into src/units.js — one module, one seam.
weather.js and config.js re-export for backward compatibility. Unit bugs now
concentrate in one module instead of leaking across three.

All 629 tests pass. Lint clean.
